styling, guard route, handle err login di client doctor

+++ client/halo-pets-doctor/src/components/AppBar.js
@@ -5,19 +5,64 @@ import {
   AppBar,
   Toolbar,
   Typography,
-  Button
+  Button,
+  Paper,
+  ClickAwayListener,
+  MenuList,
+  MenuItem,
+  Popper,
+  Grow
 } from '@material-ui/core'
 import {useHistory} from 'react-router-dom'
-
+import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 
 export default function Navbar() {
   const classes = useStyles();
   const history = useHistory()
 
-  function handleLogout() {
+  function handleLogout(event) {
+    if (anchorRef.current && anchorRef.current.contains(event.target)) {
+      return;
+    }
+
+    setOpen(false);
     localStorage.clear()
     history.push('/')
   }
+
+  // const classes = useStyles();
+  const [open, setOpen] = React.useState(false);
+  const anchorRef = React.useRef(null);
+
+  const handleToggle = () => {
+    setOpen((prevOpen) => !prevOpen);
+  };
+
+  const handleClose = (event) => {
+    if (anchorRef.current && anchorRef.current.contains(event.target)) {
+      return;
+    }
+
+    setOpen(false);
+  };
+
+  function handleListKeyDown(event) {
+    if (event.key === 'Tab') {
+      event.preventDefault();
+      setOpen(false);
+    }
+  }
+
+  // return focus to the button when we transitioned from !open -> open
+  const prevOpen = React.useRef(open);
+  React.useEffect(() => {
+    if (prevOpen.current === true && open === false) {
+      anchorRef.current.focus();
+    }
+
+    prevOpen.current = open;
+  }, [open]);
 
   return (
     <div className={classes.root}>
@@ -30,11 +75,35 @@ export default function Navbar() {
             </Typography>
           </div>
           <div>
-            <Button onClick={handleLogout}>
-              <Typography variant="h6">
-                Logout
-              </Typography>
-            </Button>
+          <Button
+            ref={anchorRef}
+            aria-controls={open ? 'menu-list-grow' : undefined}
+            aria-haspopup="true"
+            onClick={handleToggle}
+            style={{textTransform: 'none', fontSize: 16, color: 'white' }}
+          >
+            {localStorage.getItem('email')}
+            <ExpandMoreIcon />
+          </Button>
+          <Popper open={open} anchorEl={anchorRef.current} role={undefined} transition disablePortal>
+            {({ TransitionProps, placement }) => (
+              <Grow
+                {...TransitionProps}
+                style={{ transformOrigin: placement === 'bottom' ? 'center top' : 'center bottom' }}
+              >
+                <Paper>
+                  <ClickAwayListener onClickAway={handleClose}>
+                    <MenuList autoFocusItem={open} id="menu-list-grow" onKeyDown={handleListKeyDown}>
+                      <MenuItem onClick={handleLogout}><ExitToAppIcon style={{ marginRight: 8 }} /> Logout</MenuItem>
+                    </MenuList>
+                  </ClickAwayListener>
+                </Paper>
+              </Grow>
+            )}
+          </Popper>
+            {/* <Button onClick={handleLogout} size="large">
+              Logout
+            </Button> */}
           </div>
         </Toolbar>
       </AppBar>
